move seed database method
